Allow implementations to choose how to allocate memory for file objects.
This patch allows implementations of the FileOperations trait to decide
how the file objects created by the `open` function are allocated. This
allows them to choose, for example, `Arc` instead of `Box` if they need
reference couting; it also allows them to know the address of the file
object before returning it, so they can initialise any state that cannot
be moved after initialisation.

Signed-off-by: Wedson Almeida Filho <wedsonaf@google.com>