Emit `metadata` instead of `link`
It allows us to save a bit of space, ignore the duplicate object files,
and the archiving steps; e.g.:

      643544 libcompiler_builtins.rlib
    64171752 libcore.rlib

vs.

      530004 libcompiler_builtins.rmeta
    63679866 libcore.rmeta

We couldn't do it right away in [1] because `rustc` required a fix [2,3].
The fix is now in [4] and available since the 2021-01-21 nightly, so now
we can go ahead and make the change.

Fixes #75.

[1] https://github.com/Rust-for-Linux/linux/pull/52
[2] https://github.com/rust-lang/rust/issues/81117
[3] https://github.com/rust-lang/rust/pull/81118
[4] https://github.com/rust-lang/rust/commit/f9275e1092232fcb8ec117fc4acca990f57cba15

Suggested-by: bjorn3
Signed-off-by: Miguel Ojeda <ojeda@kernel.org>