Migrate to the new stack probes option
The stack probes options changed from taking a boolean to a set
of options. For the moment, use `inline-or-call` for both x86_64
and arm64, since that is what the x86 builtin targets moved to
and since we had it enabled for arm64 too (although note that
arm64 does not have it in the builtin target).

This also raises the minimum version of `rustc` to 2021-01-24,
which is just a few days from the previous one (2021-01-21), which
is easier than trying to support both formats.

Triggered by #84.

Signed-off-by: Miguel Ojeda <ojeda@kernel.org>