Samples
Reorganizes and splits tests into `samples/`

Now that we have them split, we can remove the GitHub-only ones, too,
since we have enough to test several builtins + several loadable.

However, in order to keep testing the module parameters in both
builtin vs. loadable and default vs. custom, we generate the modules
on the fly based on the base one.

Takes also the chance to update the `.config`s; as well as cleanups
the unused imports in each sample now that they are split.

Updates the docs accordingly too.

Fixes https://github.com/Rust-for-Linux/linux/issues/65.

Signed-off-by: Miguel Ojeda <ojeda@kernel.org>