Export everything as GPL
This means we are now also exporting the helpers, `core` and `alloc`
as GPL too.

It is simpler and also allows us to not worry later on if, for
instance, we want to have our own `alloc`.

In the future, if someone wants some of this as non-GPL, they can
nevertheless come to us to discuss why etc.

Signed-off-by: Miguel Ojeda <ojeda@kernel.org>