Support Rust `core::fmt::Argument` in vsprintf
This patch adds a format specifier `%pA` to `vsprintf` which formats a
pointer as `core::fmt::Arguments`. Doing so allows us to directly format
to the internal buffer of printf, so we don't have to use a temporary
buffer on the stack to pre-assemble the message on the Rust side.

This specifier is intended only to be used from Rust and not for C, so
`checkpatch.pl` is intentionally unchanged to catch any misuse.

Signed-off-by: Gary Guo <gary@garyguo.net>