rust: update `FileOperations` to use arbitrary type from `PointerWrapper`.
This allows `FileOperations` implementers to have arbitrary (wrapped)
values as `private_data` (as opposed to only wrapped `Self`).

It also allows wrappers to dictate the borrowed type. For now all
wrappers of `T` just borrow `&T`, but in a subsequent PR `Ref<T>` will
borrow to `&Ref<T>`, which allows implementations to increment the
refcount.

Signed-off-by: Wedson Almeida Filho <wedsonaf@google.com>