rust: add support for trait objects in `Ref`s.
This allows, for example, the automatic coercion from `Ref<T>` to
`Ref<dyn U>` when `U` is a trait implemented by `T`. It also allows
`Ref<dyn U>` to be used as the receiver in a trait object.

These traits are also implemented by `Arc`, and this is in preparation
for replacing all usages of `Arc` with `Ref`, which is needed before we
can remove `Arc` from the `alloc` crate.

Signed-off-by: Wedson Almeida Filho <wedsonaf@google.com>