rust: add `UniqueRef`.
This allows a ref-counted object to be allocated but initialised later.
It is useful, for example, to allocate in one context (e.g., sleepable
context) and initialise in a different context (e.g., atomic context).

It can also be used to get mutable references to the inner object after
it has been allocated, but before it can be shared (in which case it
becomes pinned and there is no safe path to the mutable inner object).

It is the last remaining feature before we can remove all usages of
`Arc<T>` from Binder.

Signed-off-by: Wedson Almeida Filho <wedsonaf@google.com>