Add `dbg!` macro from Rust standard library
The Rust standard library has a really handy macro, `dbg!`. It prints
the source location (filename and line) along with the raw source code
that is is invoked with and the `Debug` representation of the given
expression. This patch ports the macro over to the kernel crate.

For use in the kernel, we don't print with `eprintln!` (stderr),
but `pr_info!`.

Since the source code for the macro is taken from the standard library
source (with only minor adjustments), the designated file
(`std_vendor.rs`) is licensed under "Apache 2.0 OR MIT", just like its
origin. This is the same approach as with the vendored `alloc` crate.

Signed-off-by: Niklas Mohrin <dev@niklasmohrin.de>