Makefile: remove `TENTATIVE_CLANG_FLAGS`
We only really care about the target. And even if we did need
something more, it would be better to control it explicitly.

On top of that, there is now an `$(error ...)` that we would
need to avoid otherwise.

I will talk with the Clang team to see if they would be happy
moving the arch table outside so that we can reuse it.

Signed-off-by: Miguel Ojeda <ojeda@kernel.org>