rust: Add support for irqsave/irqrestore variant of spin lock/unlock.
We make the guard context an optional ulong. When locking without
disabling interrupts, we store `None` in the context; when disabling
interrupts, we store `Some(x)`, where `x` is the interrupt state before
locking. During unlock we choose which variant to use based on whether
the guard context is `None` or `Some(x)` (this check is short-circuited
by the optimiser in most cases).

Signed-off-by: Wedson Almeida Filho <wedsonaf@google.com>