rust: improve `KernelModule` initialisation.
These changes allow the init of modules to be implemented in other
modules (e.g., the main kernel crate, which isn't a separate module).
This, in turn, allows us to define module macros for specific types of
modules without procmacros (albeit with the limitation that `type` must
be the first parameter).

As an example, I redefined `module_misc_device` as a regular macro.

This is in preparation for adding macros for pci, amba, and platform
modules without the need for more procmacros.

Signed-off-by: Wedson Almeida Filho <wedsonaf@google.com>