rust: add `Revocable` type.
It allows access to objects to be safely revoked at run time.

This is useful, for example, for resources allocated during device
probe; when the device is removed, the driver should stop accessing the
device resources even if other state is kept in memory due to existing
references (i.e., device context data is ref-counted and has a non-zero
refcount after removal of the device).

Signed-off-by: Wedson Almeida Filho <wedsonaf@google.com>