ci: enable `CONFIG_WERROR`
Now that there are no warnings, we can enable `CONFIG_WERROR`.

I have a WIP to handle this and other options in the pipeline,
but I want to have this in for the moment before sending the next
patch series upstream so that we do not miss any new warnings.

Signed-off-by: Miguel Ojeda <ojeda@kernel.org>