rust: add `RevocableMutex`.
This is a mutex where access to its contents can be revoked at runtime.
This is different from `Revocable` in a few ways:
1. The caller may sleep while holding a guard.
2. Accessors are all serialised.
3. Accessing is not as cheap (because it involves acquiring the mutex).
4. The size of the object is larger (because it involves a mutex +
flag).

An example of where this a good fit to be used in device state that
holds registrations to other subsystems, for example, the PL061 driver
registers with the gpio subsystem; the registration is only used on
`probe` and is revoked on `remove`.

Signed-off-by: Wedson Almeida Filho <wedsonaf@google.com>