rust: rework Rust toolchain detection
The name of the `HAS_RUST` option has been renamed to
`RUST_IS_AVAILABLE`, and now encompasses more checks to
see whether a suitable Rust toolchain is available for
compiling the kernel.

This will allow us later to remove the `!COMPILE_TEST` bound,
which, in turn, will allow to e.g. eventually start testing
the Rust support in places that do `all{yes,mod}config`.

This also provides an easy way for users/developers to
understand why the Rust toolchain might not be available:
a Makefile target is provided which explains why that
is the case, using the same logic that Kbuild/Kconfig use;
which addresses some of the feedback we got.

Signed-off-by: Miguel Ojeda <ojeda@kernel.org>