rust: convert `platform` to use `driver`
A platform driver is now defined by implementing a trait, very much like
`amba`. There is also a module macro used that can used if a module is
just a platform driver. Lastly, drivers can specify additional data with
each supported id (to be passed back to them on `probe`).

Signed-off-by: Wedson Almeida Filho <wedsonaf@google.com>