rust: add `ThisModule` to the arguments of driver registration
So that the kernel is aware of any loadable modules hosting drivers and
can therefore increment/decrement refcounts as appropriate to prevent
unload when the driver is in use.

This should be transparent to all drivers registered with macros. Those
registered with direct calls will need to pass the extra argument.

Signed-off-by: Wedson Almeida Filho <wedsonaf@google.com>