rust: simplify file operations by removing `FileOpener`
It is currently used to allow a single `FileOperations` implementation
to be shared by multiple `FileOpener` instances. But Binder is the only
intended user I've seen so far.

So we have the extra complexity because of a single exceptional case.
This commit simplifies things for the majority: they just need to
implement `FileOperations::open`.

Signed-off-by: Wedson Almeida Filho <wedsonaf@google.com>