rust: require `Sync` and `Send` on file operations context data
This is a requirement because a shared reference to the context data can
be used in arbitrary threads (requires `Sync`) and it can be destroyed
in an arbitrary read (requires `Send`), whenever the refcount on the
file goes to zero.

We change the default to `()` so that we don't have to impose the same
requirements on `Self`.

Signed-off-by: Wedson Almeida Filho <wedsonaf@google.com>