rust: simplify lock guards by using marker types
Unify `Guard` and `GuardMut` into a single type parametrised by a
marker. The new guard implementation is the same as the existing `Guard`
except that if the marker is `WriteLock`, it also implements `DerefMut`
(which was only implemented by `GuardMut` previously).

This is in preparation for adding abstractions for read/write locks,
which means that a single lock primitive needs to implement the `Lock`
primitive in both shared and mutable modes, which is enabled by this
change as well.

Signed-off-by: Wedson Almeida Filho <wedsonaf@google.com>