rust: avoid the need of crate attributes in kernel modules
Instead of enabling the features per-module, do it globally.
This removes the need to write boilerplate such as `#![no_std]`
in kernel modules, which does not bring much value given
the allowed features are constrained anyway.

There is the question of whether to enable only the most common ones
(`allocator_api` and `global_asm`) in order to easily see uncommon
features being used. However, we only have other three features,
and they are either going to get stabilized/workarounded/removed
at some point and/or they are easy to spot anyway.

We are using `-Zcrate-attr` for convenience here, but we do not
really depend on it: like for other unstable features, if this was
the last feature to get stabilized, we would simply workaround it.
Furthermore, it is likely we will change how modules are compiled
in a way that makes this not needed.

Signed-off-by: Miguel Ojeda <ojeda@kernel.org>