rust: helpers: explain file with a top-level comment
Also copy the same `EXPORT_SYMBOL_GPL` explanation to `exports.c`.

This also sums up the discussion at https://github.com/Rust-for-Linux/linux/pull/637 about exporting or not the helpers that are not technically required at the moment (thanks @bjorn3 for the note about `rustc` not guaranteeing codegen for non-inline functions).

Signed-off-by: Miguel Ojeda <ojeda@kernel.org>