rust: add support for static synchronisation primitives
This allows drivers to declare global shared variables (`static`) that
can be mutated when the synchronisation primitive is held.

Although we want to discourage the use of globals, some drivers still
need them. For example, a driver that needs to allocate an id to each
instance of a device that is attached (and free it up when it's
detached) may have a global variable to manage ids, access to which
needs to synchronised.

We use constructors to implement this, so we select
`CONFIG_CONSTRUCTORS` automatically when `RUST` is selected.

Signed-off-by: Wedson Almeida Filho <wedsonaf@google.com>