Kbuild: add single target support for Rust (`.{o,i,s,ll}`)
`.i` maps to macro expansion, which is somewhat similar to what
is done for C and the preprocessor.

The output is automatically reformatted if `rustfmt` is installed,
which mimics what `cargo expand` does by default. The user can
override `RUSTFMT` to avoid it in the odd case is required.

We could consider adding other outputs such as `.mir`/`.hir` if they
happen to be useful in the future.

Signed-off-by: Miguel Ojeda <ojeda@kernel.org>