docs: add SVG version of Linux logo
In the Rust for Linux patch series, Jonathan Corbet requested that
we avoid binary assets (logo and favicons for the generated
documentation), possibly by recreating the logo in SVG format.

However, it turns out there is no SVG version of the Linux logo
in the kernel tree yet, only a GIF one from the pre-v2.0.0 days.
Thus this commit fixes that, which then we can use in the Rust
documentation in one way or another.

This is a vector re-illustration by Garrett LeSage of the original
logo by Larry Ewing (plus cleanups by IFo Hancroft). Garrett gave
me (privately) his permission to use it for this purpose. The
acknowledgement requirement has been made mandatory here to match
the wording of the existing one.

Link: https://lore.kernel.org/lkml/87fspk1xoc.fsf@meer.lwn.net/
Signed-off-by: Miguel Ojeda <ojeda@kernel.org>