docs: rust: replace PNG assets by single SVG logo
Now that we have a vector version of Tux, use it to re-create
in vector format the Rust for Linux one.

However, this requires a workaround since we want to provide
the logo as a local file rather than a URL. See comments
in the code.

This addresses Jonathan Corbet's comments.

Signed-off-by: Miguel Ojeda <ojeda@kernel.org>