ci: move to containers
This allows us to use different (and in particular, newer) Ubuntu images,
it is easier to move it to another CI and to use it locally by others.

This, in turn, gives us easy access to e.g. newer GCCs like 11.2 which may
spot other issues, like the one by the Intel CI the other day (which
I wanted to reproduce in the CI -- this commit includes enabling
the config option that triggered it).

It also allows us to drop `ci-bin`, since we get QEMU 6.0 from Ubuntu
and we can compile BusyBox too.

It should also avoid most networking issues since we rely only on content
from GitHub.

Signed-off-by: Miguel Ojeda <ojeda@kernel.org>