rust: ignore `-fzero-call-used-regs=%` from GCC 11
GCC 11 introduces this flag which the kernel uses and thus `bindgen`
complains, so add it to the skip list.

In the CI, so far, we are using the GCCs provided by Ubuntu 20.04,
which are prior to GCC 11.

I checked that in an alternative CI with GCC 11.2 (from a Docker Ubuntu
21.10 image) this commit fixes the problem.

Link: https://lore.kernel.org/lkml/202201310402.vCWP8CUS-lkp@intel.com/
Reported-by: kernel test robot <lkp@intel.com>
Signed-off-by: Miguel Ojeda <ojeda@kernel.org>