rust: inhibit all `bindgen` warnings in GCC builds
After requesting other CIs to start giving a go to the Rust bits,
we have got our first report. However, this also means keeping track
of the list of warnings to be skipped for `bindgen` (in the case of
GCC builds) is more involved.

Given that GCC support is very experimental (and a hack), it is
simpler to ignore all the warnings. What we need is `bindgen`
supporting GCC (or possibly that we check the results of `bindgen`
via other means).

Link: https://lore.kernel.org/lkml/CAKwvOd=7QTUH69+ZbT7e8einvgcosTbDkyohmPaUBv6_y8RfrQ@mail.gmail.com/
Suggested-by: Nick Desaulniers <ndesaulniers@google.com>
Link: https://lore.kernel.org/lkml/202201310402.vCWP8CUS-lkp@intel.com/
Reported-by: kernel test robot <lkp@intel.com>
Signed-off-by: Miguel Ojeda <ojeda@kernel.org>