rust: define `LockInfo` trait that lock "type states" must implement
This allows the definition of additional writable type states. Without
this patch, the `DerefMut` trait for guards is only implemented when the
type state is `WriteLock`; this is too restrictive for types that want
to implement more than one writable lock type state, for example, spin
locks that may have type states that have (or have not) disabled
interrupts.

This is in preparation for simplifying spinlocks, which comes in a
subsequent patch.

Signed-off-by: Wedson Almeida Filho <wedsonaf@google.com>