kbuild: support host programs written in Rust
This allows us to use Rust for any scripts used during compilation
as long as Rust is already available, instead of e.g. C, shell
scripts or Perl (Python does not seem to be required for compilation,
which would be another natural & modern option to consider).

After all, Rust is a good language for scripts too and the goal is to
have first-class support for Rust in the kernel, so one should be able
to write host programs in it too, just like with C. Not to mention
the other benefits (e.g. static typing) compared to the currently
available options. Eventually, if having a Rust toolchain ever becomes
mandatory (whether or not Rust support is optional and/or enabled),
this becomes even a better idea. Ideally, other kernel developers
will also want to start using Rust for kernel tooling, which may help
making that happen.

A downside of Rust for this use case is `rustc`'s compilation speed,
which may become concerning if e.g. a lot of programs are used.
Another downside, shared with C, is the small standard library.

I considered adding C object linking support, but it requires creating
an archive of the C objects just for linking and a bunch of other bits
which become hard to justify when we do not have a current use case
for it. Like above, if having a Rust toolchain ever becomes mandatory
for kernel compilation, then finding a use case where C code is needed
becomes much easier. Or even the other way around, we could take
advantage of Rust inside existing C programs, too.

Signed-off-by: Miguel Ojeda <ojeda@kernel.org>