rust: target: factor out and disable `emit-debug-gdb-scripts`
This option was added in `rustc` in order to disable GDB
scripts for some no-`std` targets, see
https://github.com/rust-lang/rust/issues/44993.

Therefore, even though the default in the compiler for targets
is `true`, we leave it as `false`.

If, in the future, it gets enabled, then we should take into
account the `CONFIG_GDB_SCRIPTS` option.

Signed-off-by: Miguel Ojeda <ojeda@kernel.org>