rust: target: factor out `frame-pointer` and use `may-omit` as default
With `rustc`, we can use the `-Cforce-frame-pointers` option to force
them to be emitted (so it should work like `always`). However, when
the option is set to `n`, what happens depends on the target, thus
we cannot control the option completely from the command-line.

Therefore, to have as much control as possible from the command-line,
we keep the default as `"may-omit"`, which allows us to force it
when needed.

Also includes a small comment in the `Makefile`.

Signed-off-by: Miguel Ojeda <ojeda@kernel.org>