rust: add a simple Rust mutex implementation
It doesn't have to be pinned, so its ergonomics are much better than the
one based on the C version. In particular, it doesn't require users to
use any unsafe code.

This is a temporary solution for users can live with the simpler
implementation while we don't have a better way of dealing with pinning.

Signed-off-by: Wedson Almeida Filho <wedsonaf@google.com>