rust: add `RawSpinLock`
This is needed when code sections cannot sleep even in real-time
variants of the kernel. For example, when a callback is called with a
raw spinlock held, the callee cannot use mutexes or spinlocks, they must
use a raw spinlock as well. (This happens in gpio; pl061 will be updated
to use a raw spinlock.)

Signed-off-by: Wedson Almeida Filho <wedsonaf@google.com>