rust: split `CreatableLock` into `LockFactory` and `LockIniter`
We then make `LockFactory` more generic in the sense that it is
parametrised to create a lock for any wrapped type.

This is in preparation for adding a revocable version of any
synchronisation primitive. Given for example `Mutex<T>`, we really need
a `Mutex<Inner<T>>`, which can be achieved with the new factory.

The separation of the otherwise unrelated functions also slightly
simplifies how sequence locks are implemented, only requiring `Lock` for
the struct itself, with `LockFactory` and `LockIniter` only needed in
specific functions.

Signed-off-by: Wedson Almeida Filho <wedsonaf@google.com>