rust: add `NoWaitLock`
This is a lock that never waits, that is, if it is owned by another
thread/CPU, then attempts to acquire it will fail (instead of, for
example, blocking the caller).

We'll use this to store `Waker` instances in futures and to fetch them
when reactors need to use them.

Signed-off-by: Wedson Almeida Filho <wedsonaf@google.com>