rust: support running documentation tests in-kernel
Rust has documentation tests: these are typically examples of usage
of any item (e.g. function, struct, module...). They are very convenient
because they are just written alongside the documentation. For instance:

    /// Sums two numbers.
    ///
    /// ```
    /// assert_eq!(mymod::f(10, 20), 30);
    /// ```
    pub fn f(a: i32, b: i32) -> i32 {
        a + b
    }

So far, we were compiling and running them in the host as any other
Rust documentation test. However, that meant we could not run tests
that use kernel APIs (though we were compile-testing them, which was
already useful to keep the documentation in sync with the code).

Now, the documentation tests for the `kernel` crate are transformed
into a KUnit test suite and run within the kernel at boot time,
if enabled. This means now we can run tests that use kernel APIs.

This has also other side effects, such as now using the in-kernel
`alloc`, which does not have infallible allocations (previously,
we did use the in-tree `alloc` in the built sysroot, but it was
built via `build-std` without the same `cfg` set). This requires
fixing a few tests.

Another side effect is that Clippy is now run for doctests too,
which also requires a few changes in tests.

The CI is also changed to enable KUnit, our generated test suite
and to ensure they pass.

Signed-off-by: Miguel Ojeda <ojeda@kernel.org>