rust: kernel: review doctests now that they run in the kernel
Now that doctests run within the kernel, we can perform some
changes on them to increase their utility, e.g. now unneeded
wrapper functions can be removed, and for other wrapper functions
that are still needed, we can add calls to those to actually
run the tests (hidden or not depending on whether they add
value to readers).

In addition, we enable a few more tests and reviewed other bits,
such as asserting instead of printing.

Signed-off-by: Miguel Ojeda <ojeda@kernel.org>