kallsyms: provide a bigger buffer to report the real length
With a buffer size close to `KSYM_NAME_LEN`, the "symbol too big"
report in `read_symbol` does not really provide information
to the user. Therefore, use a bigger buffer.

Furthermore, add a static assert to keep things matched.

This will be rebased/reorganized in `rust-next` as a new set
of kallsyms patches that makes more sense (instead of putting Boqun's
patch last of the three, it will go first in the series as
an improvement to the mainline kernel). This commit synchronizes
with those changes.

Fixes: d89543615548 ("script: kallsyms: Use the correct buffer size for symbols")
Signed-off-by: Miguel Ojeda <ojeda@kernel.org>