HID: uclogic: Add support for Parblo A610 PRO
The Parblo A610 PRO tablet is a UGEE v2 tablet with a bitmap dial.

The presence of the dial is set in the 7th byte of the string
descriptor, making it possible to add generic support.

Add the tablet IDs, a template for frames with dial and the required
logic to choose the right HID report descriptor template.

Signed-off-by: José Expósito <jose.exposito89@gmail.com>