docs: rust: arch-support: use simple table
reST's list tables are quite hard to read when inspecting
the source of the documents.

There are other 3 ways to create tables in reST: grids tables,
CSV tables and simple tables. Plus, of course, the "manual"
way of using a code block with custom formatting.

Use simple tables since they are lightweight and we do not need
the features of the more complex tables.

Suggested-by: Akira Yokosawa <akiyks@gmail.com>
Suggested-by: Jonathan Corbet <corbet@lwn.net>
Signed-off-by: Miguel Ojeda <ojeda@kernel.org>