rust: use Linux logo
To simplify the Rust patch series and the licensing of the logo,
let's use the Linux logo instead.

For the moment, there is no SVG version, so we use the GIF one.
When the SVG version gets merged [1], we can switch to that one instead.

This allows us to remove from the `rust` branch both the Linux SVG
logo and the Rust for Linux SVG logo, plus the `COPYING-logo` file.

Link: https://lore.kernel.org/lkml/20220528153132.8636-2-ojeda@kernel.org/ [1]
Signed-off-by: Miguel Ojeda <ojeda@kernel.org>