Re-implement concat_idents! in libmacros
This removes an unstable feature used by kernel modules. The new proc
macro also has different behavior with respect to macro hygiene. Unlike
the unstable concat_idents! macro it for example allows referring to
local variables by taking the span of the second macro as span for the
output identifier.

Signed-off-by: Björn Roy Baron <bjorn3_gh@protonmail.com>