rust: kernel: std_vendor: clarify licensing even further
Given Linus' comments on the `alloc` patch on v6, I thought it would
be a good idea to make this paragraph even more clear (i.e. that the
original code was actually under "Apache-2.0 OR MIT" -- the SPDX
identifier alone could be understood as "we decided to license this
ourselves under Apache/MIT, but it may not be the original license").

This also makes the wording more consistent with what we have for
`rust/alloc/README.md` -- they are now almost identical.

Signed-off-by: Miguel Ojeda <ojeda@kernel.org>